more cleanup and moving methods to avoid circular dep
