+++ evalml/tests/pipeline_tests/test_pipelines.py
@@ -30,13 +30,16 @@ from evalml.pipelines import (
     RegressionPipeline,
 )
 from evalml.pipelines.components import (
+    DropNullColumns,
     ElasticNetClassifier,
     Imputer,
     LogisticRegressionClassifier,
     OneHotEncoder,
     RandomForestClassifier,
     StandardScaler,
+    TargetImputer,
     Transformer,
+    Undersampler,
 )
 from evalml.pipelines.components.utils import (
     _all_estimators_used_in_search,
@@ -2313,3 +2316,123 @@ def test_get_hyperparameter_ranges():
     hyper_ranges = pipeline.get_hyperparameter_ranges(custom_hyperparameters)
 
     assert algo_ranges == hyper_ranges
+
+
+def test_make_component_dict_from_component_list():
+    assert PipelineBase._make_component_dict_from_component_list(
+        [RandomForestClassifier]
+    ) == {"Random Forest Classifier": [RandomForestClassifier, "X", "y"]}
+    assert PipelineBase._make_component_dict_from_component_list([Imputer]) == {
+        "Imputer": [Imputer, "X", "y"]
+    }
+    assert PipelineBase._make_component_dict_from_component_list(
+        [Imputer, OneHotEncoder, DropNullColumns]
+    ) == {
+        "Imputer": [Imputer, "X", "y"],
+        "One Hot Encoder": [OneHotEncoder, "Imputer.x", "y"],
+        "Drop Null Columns Transformer": [DropNullColumns, "One Hot Encoder.x", "y"],
+    }
+
+    # Test with component that modifies y (Target Imputer)
+    assert PipelineBase._make_component_dict_from_component_list(
+        [Imputer, OneHotEncoder, TargetImputer, RandomForestClassifier]
+    ) == {
+        "Imputer": [Imputer, "X", "y"],
+        "One Hot Encoder": [OneHotEncoder, "Imputer.x", "y"],
+        "Target Imputer": [TargetImputer, "One Hot Encoder.x", "y"],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "One Hot Encoder.x",
+            "Target Imputer.y",
+        ],
+    }
+
+    # Test with component that modifies X and y (Undersampler)
+    assert PipelineBase._make_component_dict_from_component_list(
+        [
+            Imputer,
+            OneHotEncoder,
+            TargetImputer,
+            DropNullColumns,
+            Undersampler,
+            RandomForestClassifier,
+        ]
+    ) == {
+        "Imputer": [Imputer, "X", "y"],
+        "One Hot Encoder": [OneHotEncoder, "Imputer.x", "y"],
+        "Target Imputer": [TargetImputer, "One Hot Encoder.x", "y"],
+        "Drop Null Columns Transformer": [
+            DropNullColumns,
+            "One Hot Encoder.x",
+            "Target Imputer.y",
+        ],
+        "Undersampler": [
+            Undersampler,
+            "Drop Null Columns Transformer.x",
+            "Target Imputer.y",
+        ],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "Undersampler.x",
+            "Undersampler.y",
+        ],
+    }
+
+    # Test with component after estimator
+    assert PipelineBase._make_component_dict_from_component_list(
+        [Imputer, RandomForestClassifier, Imputer]
+    ) == {
+        "Imputer": [Imputer, "X", "y"],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "Imputer.x",
+            "y",
+        ],
+        "Imputer_2": [Imputer, "Random Forest Classifier.x", "y"],
+    }
+
+
+def test_make_component_dict_from_component_list_with_duplicate_names():
+    assert PipelineBase._make_component_dict_from_component_list(
+        [RandomForestClassifier, RandomForestClassifier]
+    ) == {
+        "Random Forest Classifier": [RandomForestClassifier, "X", "y"],
+        "Random Forest Classifier_1": [
+            RandomForestClassifier,
+            "Random Forest Classifier.x",
+            "y",
+        ],
+    }
+    assert PipelineBase._make_component_dict_from_component_list(
+        [Imputer, Imputer, RandomForestClassifier]
+    ) == {
+        "Imputer": [Imputer, "X", "y"],
+        "Imputer_1": [Imputer, "Imputer.x", "y"],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "Imputer_1.x",
+            "y",
+        ],
+    }
+    assert PipelineBase._make_component_dict_from_component_list(
+        [TargetImputer, TargetImputer, RandomForestClassifier]
+    ) == {
+        "Target Imputer": [TargetImputer, "X", "y"],
+        "Target Imputer_1": [TargetImputer, "X", "Target Imputer.y"],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "X",
+            "Target Imputer_1.y",
+        ],
+    }
+    assert PipelineBase._make_component_dict_from_component_list(
+        [Undersampler, Undersampler, RandomForestClassifier]
+    ) == {
+        "Undersampler": [Undersampler, "X", "y"],
+        "Undersampler_1": [Undersampler, "Undersampler.x", "Undersampler.y"],
+        "Random Forest Classifier": [
+            RandomForestClassifier,
+            "Undersampler_1.x",
+            "Undersampler_1.y",
+        ],
+    }
